requirements.txt
google-generativeai==0.7.2
gradio==4.36.1
pydantic==2.7.4
numpy==1.26.4
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.3.0
rich==13.7.1
pyyaml==6.0.1
pytest==8.2.2
pytest-asyncio==0.23.7
fastapi==0.111.0
uvicorn[standard]==0.30.1
